# docs/swagger.yaml
definitions:
...
  models.ReqCreateSale:
    properties:
      about:
        type: string
      chainId:
        type: integer
      creator:
        type: string
      endTime:
        type: integer
      limitMax:
        type: string
      limitMin:
        type: string
      receiveToken:
        type: string
      saleAmount:
        type: string
      salePrice:
        type: string
      saleToken:
        type: string
      saleWay:
        description: SaleWay:general/discount
        type: string
      startTime:
        type: integer
      title:
        type: string
      whiteList:
        items:
          type: string
        type: array
    type: object
  models.ReqPurchased:
    properties:
      account:
        type: string
      buyAmount:
        type: string
      saleId:
        type: integer
    type: object
...
  models.ResCreateSale:
    properties:
      saleId:
        type: integer
      signature:
        type: string
    type: object
...
  models.ResIsWhite:
    properties:
      isWhite:
        type: boolean
    type: object
...
  models.ResPurchased:
    properties:
      signature:
        type: string
    type: object
...
  models.ResSwapList:
    properties:
      about:
        type: string
      chainId:
        type: integer
      createTime:
        type: integer
      creator:
        type: string
      endTime:
        type: integer
      limitMax:
        type: string
      limitMin:
        type: string
      originalDiscount:
        type: string
      receiveToken:
        type: string
      receiveTokenImg:
        type: string
      saleAmount:
        type: string
      saleId:
        type: integer
      salePrice:
        type: string
      saleToken:
        type: string
      saleTokenImg:
        type: string
      saleWay:
        type: string
      soldAmount:
        type: string
      startTime:
        type: integer
      status:
        type: string
      title:
        type: string
    type: object
  models.ResSwapListPage:
    properties:
      list:
        items:
          $ref: '#/definitions/models.ResSwapList'
        type: array
      total:
        type: integer
    type: object
  models.ResSwapPrices:
    properties:
      chainId:
        type: integer
      decimals:
        type: integer
      img:
        type: string
      price:
        type: number
      symbol:
        type: string
      tokenAddress:
        type: string
      tokenName:
        type: string
      updateAt:
        type: integer
      urlCoingecko:
        type: string
      urlCoinmarketcap:
        type: string
    type: object
  models.ResSwapTransactionList:
    properties:
      buy_amount:
        type: string
      buyTokenName:
        type: string
      buyer:
        type: string
      payAmount:
        type: string
      payTokenName:
        type: string
      saleId:
        type: integer
      time:
        type: integer
    type: object
  models.ResSwapTransactionListPage:
    properties:
      list:
        items:
          $ref: '#/definitions/models.ResSwapTransactionList'
        type: array
      total:
        type: integer
    type: object
...
info:
  contact: {}
paths:
...
  /stpdao/v2/swap/create:
    post:
      description: create sale
      parameters:
      - description: request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.ReqCreateSale'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ResCreateSale'
      summary: create sale
      tags:
      - swap
  /stpdao/v2/swap/isCreatorWhite:
    get:
      description: IsCreatorWhite
      parameters:
      - description: account
        in: query
        name: account
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ResIsWhite'
      summary: IsCreatorWhite
      tags:
      - swap
  /stpdao/v2/swap/isWhite:
    get:
      description: isWhite
      parameters:
      - description: account
        in: query
        name: account
        required: true
        type: string
      - description: saleId
        in: query
        name: saleId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ResIsWhite'
      summary: isWhite
      tags:
      - swap
  /stpdao/v2/swap/list:
    get:
      description: sale list
      parameters:
      - description: 'status: soon/normal/ended'
        in: query
        name: status
        type: string
      - description: saleId
        in: query
        name: saleId
        type: integer
      - description: offset,page
        in: query
        name: offset
        required: true
        type: integer
      - description: limit,page
        in: query
        name: limit
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ResSwapListPage'
      summary: sale list
      tags:
      - swap
  /stpdao/v2/swap/prices:
    get:
      description: prices
      parameters:
      - description: chainId
        in: query
        name: chainId
        required: true
        type: integer
      - description: tokens, separate by comma, if nil, return all
        in: query
        name: tokens
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ResSwapPrices'
      summary: prices
      tags:
      - swap
  /stpdao/v2/swap/purchased:
    post:
      description: purchased sale
      parameters:
      - description: request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.ReqPurchased'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ResPurchased'
      summary: purchased sale
      tags:
      - swap
  /stpdao/v2/swap/transactions:
    get:
      description: transactions list
      parameters:
      - description: saleId
        in: query
        name: saleId
        required: true
        type: integer
      - description: offset,page
        in: query
        name: offset
        required: true
        type: integer
      - description: limit,page
        in: query
        name: limit
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ResSwapTransactionListPage'
      summary: transactions list
      tags:
      - swap
...
swagger: "2.0"